+++ src/runner/PlutoRunner/src/display/Exception.jl
@@ -1,0 +1,118 @@
+
+"Downstream packages can set this to false to obtain unprettified stack traces."
+const PRETTY_STACKTRACES = Ref(true)
+
+
+
+function frame_is_from_plutorunner(frame::Base.StackTraces.StackFrame)
+    if frame.linfo isa Core.MethodInstance
+        frame.linfo.def.module === PlutoRunner
+    else
+        endswith(String(frame.file), "PlutoRunner.jl")
+    end
+end
+
+frame_is_from_usercode(frame::Base.StackTraces.StackFrame) = occursin("#==#", String(frame.file))
+
+function frame_url(frame::Base.StackTraces.StackFrame)
+    if frame.linfo isa Core.MethodInstance
+        Base.url(frame.linfo.def)
+    elseif frame.linfo isa Method
+        Base.url(frame.linfo)
+    else
+        nothing
+    end
+end
+
+function format_output(val::CapturedException; context=default_iocontext)
+    if has_julia_syntax && val.ex isa PrettySyntaxError
+        dict = convert_parse_error_to_dict(val.ex.ex.detail)
+        return dict, MIME"application/vnd.pluto.parseerror+object"()
+    end
+
+    stacktrace = if PRETTY_STACKTRACES[]
+        ## We hide the part of the stacktrace that belongs to Pluto's evalling of user code.
+        stack = [s for (s, _) in val.processed_bt]
+
+        # function_wrap_index = findfirst(f -> occursin("function_wrapped_cell", String(f.func)), stack)
+
+        function_wrap_index = findlast(frame_is_from_usercode, stack)
+        internal_index = findfirst(frame_is_from_plutorunner, stack)
+        
+        limit = if function_wrap_index !== nothing
+            function_wrap_index
+        elseif internal_index !== nothing
+            internal_index - 1
+        else
+            nothing
+        end
+        stack_relevant = stack[1:something(limit, end)]
+
+        pretty = map(stack_relevant) do s
+            Dict(
+                :call => pretty_stackcall(s, s.linfo),
+                :inlined => s.inlined,
+                :from_c => s.from_c,
+                :file => basename(String(s.file)),
+                :path => String(s.file),
+                :line => s.line,
+                :url => frame_url(s),
+                :linfo_type => string(typeof(s.linfo)),
+            )
+        end
+    else
+        val
+    end
+
+    Dict{Symbol,Any}(:msg => sprint(try_showerror, val.ex), :stacktrace => stacktrace), MIME"application/vnd.pluto.stacktrace+object"()
+end
+
+
+# from the Julia source code:
+function pretty_stackcall(frame::Base.StackFrame, linfo::Nothing)::String
+    if frame.func isa Symbol
+        if occursin("function_wrapped_cell", String(frame.func))
+            "top-level scope"
+        else
+            String(frame.func)
+        end
+    else
+        repr(frame.func)
+    end
+end
+
+function pretty_stackcall(frame::Base.StackFrame, linfo::Core.CodeInfo)
+    "top-level scope"
+end
+
+function pretty_stackcall(frame::Base.StackFrame, linfo::Core.MethodInstance)
+    if linfo.def isa Method
+        @static if isdefined(Base.StackTraces, :show_spec_linfo) && hasmethod(Base.StackTraces.show_spec_linfo, Tuple{IO, Base.StackFrame})
+            sprint(Base.StackTraces.show_spec_linfo, frame; context=:backtrace => true)
+
+        else
+            split(string(frame), " at ") |> first
+        end
+    else
+        sprint(Base.show, linfo)
+    end
+end
+
+function pretty_stackcall(frame::Base.StackFrame, linfo::Method)
+    sprint(Base.show_tuple_as_call, linfo.name, linfo.sig)
+end
+
+function pretty_stackcall(frame::Base.StackFrame, linfo::Module)
+    sprint(Base.show, linfo)
+end
+
+
+"Because even showerror can error... 👀"
+function try_showerror(io::IO, e, args...)
+    try
+        showerror(io, e, args...)
+    catch show_ex
+        print(io, "\nFailed to show error:\n\n")
+        try_showerror(io, show_ex, stacktrace(catch_backtrace()))
+    end
+end
